requirements.txt
backcall==0.1.0
decorator==4.4.0
Django==2.2
django-sql-explorer==1.1.2
djangorestframework==3.9.2
gunicorn==19.9.0
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
numpy==1.16.3
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
whitenoise==4.1.2
